+++ src/SecondWeek/homework1/entity/SummerHouse.java
@@ -1,7 +1,7 @@
 package SecondWeek.homework1.entity;
 
 public class SummerHouse extends Housing {
-    public SummerHouse(Double price, int squareMeter, int numberOfRooms, int numberOfLivingRooms) {
+    public SummerHouse(Long price, int squareMeter, int numberOfRooms, int numberOfLivingRooms) {
         super(price, squareMeter, numberOfRooms, numberOfLivingRooms);
     }
