+++ src/SecondWeek/homework1/entity/Villa.java
@@ -1,7 +1,7 @@
 package SecondWeek.homework1.entity;
 
 public class Villa extends Housing {
-    public Villa(Double price, int squareMeter, int numberOfRooms, int numberOfLivingRooms) {
+    public Villa(Long price, int squareMeter, int numberOfRooms, int numberOfLivingRooms) {
         super(price, squareMeter, numberOfRooms, numberOfLivingRooms);
     }
